fix: 날씨 api 수정 #S07P31A703-251

+++ backend/community/src/main/java/com/a703/community/controller/WeatherController.java
@@ -19,8 +19,8 @@ public class WeatherController {
     private final WeatherUtil weatherUtil;
 
     @GetMapping
-    public ResponseEntity<WeatherDto> weather( @RequestParam(name = "lat") int lat,
-                                               @RequestParam(name = "lng") int lng) throws IOException {
+    public ResponseEntity<WeatherDto> weather( @RequestParam(name = "lat") double lat,
+                                               @RequestParam(name = "lng") double lng) throws IOException {
         WeatherDto weatherDto = weatherUtil.lookUpWeather(lat,lng);
         return ResponseEntity.ok().body(weatherDto);
